PHAIN-133: Generate OpenAPI schema via CLI
This change adds a `make generate-openapi-spec` command to generate an OpenAPI schema JSON file.
We have been using this process within the PHA API for some time.

To avoid needing to create a bunch of stub config and other misdirections during the application startup, a new startup specifically for Swagger has been added which instantiates just enough for Swagger to be able to generate the document.