PHAIN-133: Specify the .net version for Swagger to generate from and roll back the .NET 8 version to earlier
In a Windows environment a wildcard in the path is not supported so the `dotnet swagger` command fails.

If a developer has an older version of .NET 8 installed the `global.json` file is ignored, this rolls it back to the earliest/early version of .NET 8 and allows minor incremental changes.