PHAIN-133: Generate OpenAPI schema (#141)
* PHAIN-133: Generate OpenAPI schema via CLI

This change adds a `make generate-openapi-spec` command to generate an OpenAPI schema JSON file.
We have been using this process within the PHA API for some time.

To avoid needing to create a bunch of stub config and other misdirections during the application startup, a new startup specifically for Swagger has been added which instantiates just enough for Swagger to be able to generate the document.

* PHAIN-133: Specify the .net version for Swagger to generate from and roll back the .NET 8 version to earlier

In a Windows environment a wildcard in the path is not supported so the `dotnet swagger` command fails.

If a developer has an older version of .NET 8 installed the `global.json` file is ignored, this rolls it back to the earliest/early version of .NET 8 and allows minor incremental changes.

* PHAIN-133: Ignore private fields

Fields that start with an _ are private and should not be exposed in the OpenAPI schema.